chore: handle exception of delete board

+++ src/main/kotlin/plus/voyage/framework/service/BoardService.kt
@@ -2,7 +2,6 @@ package plus.voyage.framework.service
 
 import jakarta.transaction.Transactional
 import org.springframework.security.access.AccessDeniedException
-import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.stereotype.Service
 import plus.voyage.framework.dto.*
 import plus.voyage.framework.entity.Board
@@ -31,9 +30,9 @@ class BoardService(
     }
 
     fun getAll(): BoardListResponse {
-        val username = SecurityContextHolder.getContext().authentication.name
+        val currentUser = userService.getCurrentUser()
         val boards = boardRepository.findAll()
-            .map { BoardItem.from(it, username) }
+            .map { BoardItem.from(it, currentUser.username) }
 
         return BoardListResponse(
             totalCounts = boards.size,
@@ -42,12 +41,12 @@ class BoardService(
     }
 
     fun getById(boardId: Int): BoardItem {
-        val username = SecurityContextHolder.getContext().authentication.name
+        val currentUser = userService.getCurrentUser()
         val board: Board = boardRepository.findById(boardId).orElseThrow {
             throw BoardNotFoundException("$boardId 번 게시글을 찾을 수 없습니다.")
         }
 
-        return BoardItem.from(board, username)
+        return BoardItem.from(board, currentUser.username)
     }
 
     @Transactional
@@ -69,6 +68,14 @@ class BoardService(
 
     @Transactional
     fun delete(boardId: Int) {
+        val board: Board = boardRepository.findById(boardId).orElseThrow {
+            throw BoardNotFoundException("$boardId 번 게시글을 찾을 수 없습니다.")
+        }
+        val currentUser = userService.getCurrentUser()
+        if (currentUser.username != board.user.username && currentUser.role != Role.ADMIN) {
+            throw AccessDeniedException("게시글 삭제 권한이 없습니다.")
+        }
+
         boardRepository.deleteById(boardId)
     }
 }
